Move action/store handling out of CampaignForm
This is necessary not only to be able to pass a (filtered) list of actions to
the component, but also for future sharing of the CampaignForm between this
app and a separate caller app.
